refactor(marketplace): extract hub catalog model
